Multiple changes
I have made changes so that when you click play on the website for number recognition, you actually play the nrplaygame.php file. I have also made it so that the won_game.php file redirects to the welcome page after 3 seconds. I have changed my password. Lastly, in right_answer.php I have made it so that the username gets carried across pages so that when the redirect occurs after a game has been won, the username is kept.